requirements.txt
# google-generativeai langchain-google-genai
# pip install langchain python-dotenv

# [[source]]
# url = "https://pypi.org/simple"
# verify_ssl = true
# name = "pypi"
langchain
openai
langchain-google-genai
chromadb
python-dotenv 
# [packages]
# langchain = "==0.0.352"
# openai = "==0.27.8"
# langchain-google-genai
# python-dotenv = "==1.0.0"
 
# [dev-packages]
 
# [requires]
# python_version = "3.11"

